Remove comment depth limit. #430
This was an arbitrary limit that doesn't make sense alongside other proposed changes.
It is currently limiting usability of Golos, so we are removing it entirely.

This commit is import of steemit#767 .

+++ libraries/chain/include/golos/chain/comment_object.hpp
@@ -127,7 +127,7 @@ namespace golos {
             time_point_sec active; ///< the last time this post was "touched" by voting or reply
             time_point_sec last_payout;
 
-            uint8_t depth = 0; ///< used to track max nested depth
+            uint16_t depth = 0; ///< used to track max nested depth
             uint32_t children = 0; ///< used to track the total number of children, grandchildren, etc...
 
             /**
